rename main.js -> when.js, fix exports
