+++ test/when.js
